HotelReservationTest.java
package com.capgemini.assignment.hotelreservation;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;

import org.junit.Test;

public class HotelReservationTest {
	@Test
	public void addHotelForRegularCustomerTest() {
		HotelReservation hotelReservation = new HotelReservation();
		hotelReservation.addHotel("Lakewood", 3, 110, 90, 80, 80);
		hotelReservation.addHotel("Bridgewood", 4, 150, 50, 110, 50);
		hotelReservation.addHotel("Ridgewood", 5, 220, 150, 100, 40);
		int size = hotelReservation.getHotelList().size();
		assertEquals(3, size);		
	}
	
	@Test
	public void findCheapestAndHighestRatingHotelForRewardCustomerTest() throws Exception {
		HotelReservation hotelReservation = new HotelReservation();
		hotelReservation.addHotel("Lakewood", 3, 110, 90, 80, 80);
		hotelReservation.addHotel("Bridgewood", 4, 150, 50, 110, 50);
		hotelReservation.addHotel("Ridgewood", 5, 220, 150, 100, 40);
		String output = hotelReservation.findCheapestHotel("reward, 11Sep2020, 12Sep2020");
		boolean check = output.equals("Ridgewood, Rating: 5 and Total Rates: $140");
		assertTrue(check);
	}
}
